Build fails on aarch64
[Issue]    http://web.sec.samsung.net/bugzilla/show_bug.cgi?id=15312
[Problem]  Incompatible types used in 32 and 64 bit architecture
[Solution] Use proper types
[Verify]   Build and run browser in both architectures.

To build aarch64 uncomment proper line in build_mobile.sh

+++ services/QuickAccess/QuickAccess.h
@@ -19,6 +19,7 @@
 
 #include <Evas.h>
 #include <boost/signals2/signal.hpp>
+#include <cstdint>
 
 #include "AbstractUIComponent.h"
 #include "AbstractService.h"
@@ -75,7 +76,7 @@ private:
 
 #if PROFILE_MOBILE
     void addBookmarkManagerTile();
-    void setIndexPage(const int page) const;
+    void setIndexPage(const uintptr_t page) const;
 #else
     Evas_Object* createBottomButton(Evas_Object *parent);
 #endif
